Rename tensorflow package
