+++ aws-predictions-tensorflow/src/main/AndroidManifest.xml
@@ -14,4 +14,4 @@
    permissions and limitations under the License.
 -->
 
-<manifest package="com.amplifyframework.predictions" />
+<manifest package="com.amplifyframework.predictions.tensorflow" />
