+++ aws-predictions-tensorflow/src/main/java/com/amplifyframework/predictions/tensorflow/operation/TFLiteInterpretOperation.java
@@ -13,15 +13,16 @@
  * permissions and limitations under the License.
  */
 
-package com.amplifyframework.predictions.operation;
+package com.amplifyframework.predictions.tensorflow.operation;
 
 import androidx.annotation.NonNull;
 
 import com.amplifyframework.core.Consumer;
 import com.amplifyframework.predictions.PredictionsException;
-import com.amplifyframework.predictions.request.TFLiteTextClassificationRequest;
+import com.amplifyframework.predictions.operation.InterpretOperation;
+import com.amplifyframework.predictions.tensorflow.request.TFLiteTextClassificationRequest;
 import com.amplifyframework.predictions.result.InterpretResult;
-import com.amplifyframework.predictions.service.TFLitePredictionsService;
+import com.amplifyframework.predictions.tensorflow.service.TFLitePredictionsService;
 
 import java.util.Objects;
 import java.util.concurrent.ExecutorService;
